user_service/requirements.txt
asgiref
Django
djangorestframework
psycopg
psycopg2-binary
python-dotenv
sqlparse
typing_extensions
